requirements.txt
# requirements.txt
telethon
pytesseract
Pillow
pydantic-settings
python-dotenv
colorama

# Dependência do Banco de Dados
sqlalchemy